+++ packages/playground-react/pages/table/constant/icon.tsx
@@ -1,0 +1,16 @@
+export const SortingIcon = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="17"
+    height="14"
+    viewBox="0 0 17 14"
+  >
+    <path
+      id="다각형_3"
+      data-name="다각형 3"
+      d="M7.645,1.408a1,1,0,0,1,1.71,0l6.723,11.073A1,1,0,0,1,15.223,14H1.777a1,1,0,0,1-.855-1.519Z"
+      transform="translate(17 14) rotate(180)"
+      fill="#007aff"
+    />
+  </svg>
+);
