+++ packages/playground-react/pages/table/constant/mock.ts
@@ -3,45 +3,28 @@ export const noticeList = [
     id: 1,
     title: '스터디 가입전 해당 공지사항을 숙지해주세요.',
     writer: '파크',
-    date: '2022.10.22',
+    date: 1669750385300,
     view: 101,
   },
   {
     id: 2,
     title: '스터디 전체 공지 드립니다.',
     writer: '도니',
-    date: '2022.09.20',
+    date: 1662476400000,
     view: 110,
   },
   {
     id: 3,
     title: '스터디 전체 공지 드립니다.',
     writer: '햄디',
-    date: '2022.09.20',
+    date: 1662476400000,
     view: 9,
   },
   {
     id: 4,
     title: '스터디 전체 공지 드립니다.',
     writer: '제이미',
-    date: '2022.09.22',
+    date: 1668750300553,
     view: 9,
   },
 ];
-
-export const SortingIcon = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="17"
-    height="14"
-    viewBox="0 0 17 14"
-  >
-    <path
-      id="다각형_3"
-      data-name="다각형 3"
-      d="M7.645,1.408a1,1,0,0,1,1.71,0l6.723,11.073A1,1,0,0,1,15.223,14H1.777a1,1,0,0,1-.855-1.519Z"
-      transform="translate(17 14) rotate(180)"
-      fill="#007aff"
-    />
-  </svg>
-);
